# frontend/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    AbstractTrackPart:
      oneOf:
      - $ref: "#/components/schemas/BlockStraight"
      - $ref: "#/components/schemas/Straight"
      - $ref: "#/components/schemas/Curve"
      - $ref: "#/components/schemas/Signal"
      - $ref: "#/components/schemas/Turnout"
      - $ref: "#/components/schemas/Uncoupler"
      discriminator:
        propertyName: trackPartType
      type: object
      properties:
        id:
          type: integer
          format: int64
        gridPosition:
          $ref: "#/components/schemas/GridPosition"
        trackPartType:
          type: string
        rotationAngle:
          type: number
          format: double
      required:
      - trackPartType
    Add:
      type: object
      required:
      - trackPart
      properties:
        trackPart:
          $ref: "#/components/schemas/AbstractTrackPart"
...
    BlockStraight:
      allOf:
      - $ref: "#/components/schemas/Straight"
      description: type for a track part
      type: object
      properties:
        blockLength:
          type: integer
          format: int32
        leftTrackBlock:
          $ref: "#/components/schemas/TrackBlock"
        middleTrackBlock:
          $ref: "#/components/schemas/TrackBlock"
        rightTrackBlock:
          $ref: "#/components/schemas/TrackBlock"
        endGridPosition:
          $ref: "#/components/schemas/GridPosition"
        gridLength:
          type: integer
          format: int32
        allTrackBlocks:
          type: array
          items:
            $ref: "#/components/schemas/TrackBlock"
...
    BusDataConfiguration:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bus:
          type: integer
          format: int32
        address:
          type: integer
          format: int32
        bit:
          type: integer
          format: int32
        bitState:
          type: boolean
...
    ChangeTrackDto:
      type: object
      properties:
        addActions:
          type: array
          items:
            $ref: "#/components/schemas/Add"
        moveActions:
          type: array
          items:
            $ref: "#/components/schemas/Move"
        rotateActions:
          type: array
          items:
            $ref: "#/components/schemas/Rotate"
        empty:
          type: boolean
...
    Curve:
      description: type for a track part
      type: object
      properties:
        id:
          type: integer
          format: int64
        gridPosition:
          $ref: "#/components/schemas/GridPosition"
        trackPartType:
          type: string
        direction:
          $ref: "#/components/schemas/DIRECTION1"
        nextGridPosition:
          $ref: "#/components/schemas/GridPosition"
        lastGridPosition:
          $ref: "#/components/schemas/GridPosition"
        rotationAngle:
          type: number
          format: double
      required:
      - trackPartType
...
    DIRECTION:
      type: string
      enum:
      - HORIZONTAL
      - VERTICAL
    DIRECTION1:
      type: string
      enum:
      - BOTTOM_RIGHT
      - BOTTOM_LEFT
      - TOP_LEFT
      - TOP_RIGHT
    DIRECTION2:
      type: string
      enum:
      - RIGHT
      - LEFT
...
    DRIVING_LEVEL_ADJUST_TYPE:
      type: string
      enum:
      - NONE
      - ENTER
      - EXIT
...
    EventConfiguration:
      type: object
      properties:
        id:
          type: integer
          format: int64
        stateOnConfig:
          $ref: "#/components/schemas/BusDataConfiguration"
        stateOffConfig:
          $ref: "#/components/schemas/BusDataConfiguration"
    FUNCTION:
      type: string
      enum:
      - HP0
      - HP1
      - HP2
      - HP0_SH1
    GridPosition:
      type: object
      properties:
        id:
          type: integer
          format: int64
        x:
          type: integer
          format: int32
        "y":
          type: integer
          format: int32
...
    Move:
      type: object
      required:
      - trackPartId
      - gridPosition
      properties:
        trackPartId:
          type: integer
          format: int64
        gridPosition:
          $ref: "#/components/schemas/GridPosition"
...
    PRESENTATION:
      type: string
      enum:
      - LEFT_TO_RIGHT
      - RIGHT_TO_LEFT
      - BOTTOM_TO_TOP
      - TOP_TO_BOTTOM
...
    Rotate:
      type: object
      required:
      - trackPartId
      properties:
        trackPartId:
          type: integer
          format: int64
        value:
          type: string
...
    SIGNAL_TYPE:
      type: string
      enum:
      - BLOCK
      - ENTER
      - EXIT
      - BEFORE
...
    Signal:
      allOf:
      - $ref: "#/components/schemas/Straight"
      description: type for a track part
      type: object
      properties:
        type:
          $ref: "#/components/schemas/SIGNAL_TYPE"
        signalConfigRed1:
          $ref: "#/components/schemas/BusDataConfiguration"
        signalConfigRed2:
          $ref: "#/components/schemas/BusDataConfiguration"
        signalConfigGreen1:
          $ref: "#/components/schemas/BusDataConfiguration"
        signalConfigGreen2:
          $ref: "#/components/schemas/BusDataConfiguration"
        signalConfigYellow1:
          $ref: "#/components/schemas/BusDataConfiguration"
        signalConfigYellow2:
          $ref: "#/components/schemas/BusDataConfiguration"
        signalConfigWhite:
          $ref: "#/components/schemas/BusDataConfiguration"
        stopBlock:
          $ref: "#/components/schemas/TrackBlock"
...
    Straight:
      description: type for a track part
      type: object
      properties:
        id:
          type: integer
          format: int64
        gridPosition:
          $ref: "#/components/schemas/GridPosition"
        trackPartType:
          type: string
        direction:
          $ref: "#/components/schemas/DIRECTION"
      required:
      - trackPartType
...
    TrackBlock:
      type: object
      properties:
        id:
          type: integer
          format: int64
        blockFunction:
          $ref: "#/components/schemas/BusDataConfiguration"
        name:
          type: string
        forwardTargetDrivingLevel:
          type: integer
          format: int32
        backwardTargetDrivingLevel:
          type: integer
          format: int32
        drivingLevelAdjustType:
          $ref: "#/components/schemas/DRIVING_LEVEL_ADJUST_TYPE"
        feedback:
          type: boolean
        displayValue:
          type: string
...
    Turnout:
      description: type for a track part
      type: object
      properties:
        id:
          type: integer
          format: int64
        gridPosition:
          $ref: "#/components/schemas/GridPosition"
        trackPartType:
          type: string
        currentDirection:
          $ref: "#/components/schemas/DIRECTION2"
        currentPresentation:
          $ref: "#/components/schemas/PRESENTATION"
        toggleFunction:
          $ref: "#/components/schemas/BusDataConfiguration"
        eventConfiguration:
          $ref: "#/components/schemas/EventConfiguration"
        nextGridPositionForStateBranchToLeft:
          $ref: "#/components/schemas/GridPosition"
        nextGridPositionForStateBranchToRight:
          $ref: "#/components/schemas/GridPosition"
      required:
      - trackPartType
    Uncoupler:
      allOf:
      - $ref: "#/components/schemas/Straight"
      description: type for a track part
      type: object
      properties:
        toggleFunction:
          $ref: "#/components/schemas/BusDataConfiguration"
        eventConfiguration:
          $ref: "#/components/schemas/EventConfiguration"
...
paths:
...
  /api/track:
    get:
      tags:
      - track
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AbstractTrackPart"
      summary: List All
    put:
      tags:
      - track
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeTrackDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Change
  /api/track-block:
    get:
      tags:
      - track
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackBlock"
      summary: List All
    post:
      tags:
      - track
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackBlock"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackBlock"
          description: Created
        "400":
          description: Bad Request
      summary: Create
  /api/track-block/{id}:
    put:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackBlock"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Update
    get:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackBlock"
      summary: Get By Id
    delete:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Delete
  /api/track/{id}:
    put:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbstractTrackPart"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Update
    get:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbstractTrackPart"
      summary: Load Track Part
    delete:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Delete
  /api/track/{id}/switch-signal:
    post:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/FUNCTION"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Switch Signal
  /api/track/{id}/toggle:
    post:
      tags:
      - track
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Toggle Track Part
...
info:
  title: controlcenter API
  version: 0.3-SNAPSHOT